backend/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.0.0
deprecation==2.1.0
distro==1.8.0
docstring-parser==0.15
exceptiongroup==1.2.0
fastapi==0.109.2
ffmpeg-python==0.2.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.1
gotrue==2.1.0
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.0
instructor==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
modal==0.56.4353
multidict==6.0.4
openai==1.4.0
packaging==23.2
postgrest==0.13.0
protobuf==4.25.1
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.2
requests==2.31.0
rich==13.7.0
sigtools==4.0.1
sniffio==1.3.0
starlette==0.36.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.1
synchronicity==0.5.3
tblib==3.0.0
toml==0.10.2
tqdm==4.66.1
twilio==8.11.0
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.7
typing==3.7.4.3
typing_extensions==4.9.0
urllib3==2.1.0
watchfiles==0.21.0
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.17.0
